Get daily set of curated items
